geralt-ai-fastapi/requirements.txt
# =============================================================================
# GeraltAI API - Requirements
# =============================================================================

# -----------------------------------------------------------------------------
# FastAPI Core
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# -----------------------------------------------------------------------------
# Database Clients
# -----------------------------------------------------------------------------
pymongo>=4.0.0
redis>=5.0.0
elasticsearch[async]>=8.0.0,<9.0.0
pymilvus>=2.3.0
minio>=7.0.0

# -----------------------------------------------------------------------------
# AI/ML Providers
# -----------------------------------------------------------------------------
# Gemini (NEW)
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5

# OpenAI
openai>=1.0.0

# Mistral
mistralai

# Cohere (Reranking)
cohere>=4.0.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# -----------------------------------------------------------------------------
# Document Processing
# -----------------------------------------------------------------------------
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0
xlrd>=2.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
chardet>=5.0.0
pdfkit>=1.0.0

# -----------------------------------------------------------------------------
# Media Processing
# -----------------------------------------------------------------------------
pydub>=0.25.0
SpeechRecognition>=3.10.0
moviepy==1.0.3
pytube>=15.0.0
yt-dlp>=2023.12.0
imageio[ffmpeg]>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10

# -----------------------------------------------------------------------------
# Authentication
# -----------------------------------------------------------------------------
msal>=1.24.0
argon2-cffi>=23.1.0

# -----------------------------------------------------------------------------
# Async Tasks
# -----------------------------------------------------------------------------
celery>=5.3.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
colorama>=0.4.6
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
watchdog>=3.0.0

# -----------------------------------------------------------------------------
# Web Automation (if needed)
# -----------------------------------------------------------------------------
selenium>=4.15.0

# -----------------------------------------------------------------------------
# Realtime
# -----------------------------------------------------------------------------
python-socketio>=5.0.0

# -----------------------------------------------------------------------------
# Production Server
# -----------------------------------------------------------------------------
gunicorn>=21.0.0

# -----------------------------------------------------------------------------
# Security
# -----------------------------------------------------------------------------
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0  # passlib 1.7.4 is incompatible with bcrypt 5.x